pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sysmon"
version = "0.1.0"
description = "系統查詢工具 - 網路/系統資訊查詢平台"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.41.0",
    "psutil>=6.0.0",
    "requests>=2.32.0",
    "dnspython>=2.6.0",
    "python-whois>=0.9.4",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pyopenssl>=24.0.0",
    "cryptography>=42.0.0",
    "user-agents>=2.2.0",
    "plotly>=5.22.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "validators>=0.34.0",
    "ipwhois>=1.3.0",
]

[project.scripts]
sysmon = "sysmon.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["sysmon"]
